[리팩토링] 클래스 분리  Input, View 단위로 분리 후 Input에 재시작 버튼 상수로 추가
